# Add go-kit logger benchmarks