--- a/Makefile
+++ b/Makefile
@@ -1,7 +1,7 @@
 GOTEST_FLAGS=-cpu=1,2,4 -benchmem -benchtime=5s
 
-TEXT_PKGS=Logrus Log15 Gologging Seelog
-JSON_PKGS=Logrus Log15
+TEXT_PKGS=Gokit Logrus Log15 Gologging Seelog
+JSON_PKGS=Gokit Logrus Log15
 
 TEXT_PKG_TARGETS=$(addprefix test-text-,$(TEXT_PKGS))
 JSON_PKG_TARGETS=$(addprefix test-json-,$(JSON_PKGS))
@@ -15,6 +15,7 @@ deps:
 	go get -u gopkg.in/inconshreveable/log15.v2
 	go get -u github.com/op/go-logging
 	go get -u github.com/cihub/seelog
+	go get -u github.com/go-kit/kit/log
 
 test: test-text test-json