@imkira Another logger for your benchmarks! Please wait until I get a thumbs-up from @ChrisHines before merging, I want to validate the approach :)

@ChrisHines this was my best guess at a fair way to do the comparison. Does it look reasonable?

```
$ for format in JSON Text ; for posneg in Positive Negative ; go test -benchmem -bench=.$format$posneg ; end ; end

BenchmarkGokitJSONPositive-4      300000              4804 ns/op            1328 B/op         36 allocs/op
BenchmarkLog15JSONPositive-4      200000             11835 ns/op            1800 B/op         40 allocs/op
BenchmarkLogrusJSONPositive-4     300000              5853 ns/op            2345 B/op         43 allocs/op
ok      github.com/peterbourgon/go-loggers-bench        5.804s

BenchmarkGokitJSONNegative-4     5000000               328 ns/op             232 B/op          9 allocs/op
BenchmarkLog15JSONNegative-4     2000000               895 ns/op             392 B/op          9 allocs/op
BenchmarkLogrusJSONNegative-4    1000000              1015 ns/op             832 B/op         10 allocs/op
ok      github.com/peterbourgon/go-loggers-bench        5.724s

BenchmarkGokitTextPositive-4             1000000              1406 ns/op             352 B/op          7 allocs/op
BenchmarkGologgingTextPositive-4         1000000              2021 ns/op             952 B/op         17 allocs/op
BenchmarkLog15TextPositive-4              200000              7688 ns/op            1120 B/op         24 allocs/op
BenchmarkLogrusTextPositive-4            1000000              2679 ns/op             880 B/op         17 allocs/op
BenchmarkSeelogTextPositive-4             300000              4326 ns/op             456 B/op         12 allocs/op
ok      github.com/peterbourgon/go-loggers-bench        10.168s

BenchmarkGokitTextNegative-4            20000000               119 ns/op              64 B/op          3 allocs/op
BenchmarkGologgingTextNegative-4        10000000               210 ns/op             160 B/op          3 allocs/op
BenchmarkLog15TextNegative-4             3000000               510 ns/op             128 B/op          1 allocs/op
BenchmarkLogrusTextNegative-4           50000000                45.3 ns/op            16 B/op          1 allocs/op
BenchmarkSeelogTextNegative-4           10000000               116 ns/op              64 B/op          3 allocs/op
ok      github.com/peterbourgon/go-loggers-bench        10.513s
```