@imkira 2016-05-13
@peterbourgon 
Thanks.

If/when you get the thumbs up, could you also update README for all entries on your machine?