@ChrisHines 2016-05-13
@peterbourgon The overall approach looks reasonable. It is not the only approach, but it is probably the approach someone who cares about performance, and doesn't want to wrap all their logging calls with a separate `if` statement, would appreciate the most.

To be fair, though, we need to wrap the `io.Writer` with a layer to make it safe for concurrent use. As you know I'm working on adding a type to Go kit to do just that, but I haven't submitted the PR yet. It's up to you whether you want to wait for that to get merged to Go kit or add an implementation directly to this PR. It will be interesting to see how much that changes the performance.