@peterbourgon 2016-05-13
> To be fair, though, we need to wrap the io.Writer with a layer to make it safe for concurrent use . . . It will be interesting to see how much that changes the performance.

Not much difference. With this patch

```patch
diff --git a/gokit_test.go b/gokit_test.go
index 0b27691..331064c 100644
--- a/gokit_test.go
+++ b/gokit_test.go
@@ -1,6 +1,7 @@
 package bench
 
 import (
+	"sync"
 	"testing"
 
 	"github.com/go-kit/kit/log"
@@ -26,8 +27,21 @@ type leveledLogger struct {
 	Error log.Logger
 }
 
+// For now, manually synchronize writes to the stream.
+type synchronizedStream struct {
+	mtx sync.Mutex
+	blackholeStream
+}
+
+func (s *synchronizedStream) Write(p []byte) (int, error) {
+	s.mtx.Lock()
+	n, err := s.blackholeStream.Write(p)
+	s.mtx.Unlock()
+	return n, err
+}
+
 func BenchmarkGokitJSONPositive(b *testing.B) {
-	stream := &blackholeStream{}
+	stream := &synchronizedStream{}
 	logger := log.NewContext(log.NewJSONLogger(stream)).With("ts", log.DefaultTimestampUTC)
 	lvllog := newLeveledLogger(logger, true)
 
@@ -45,7 +59,7 @@ func BenchmarkGokitJSONPositive(b *testing.B) {
 }
 
 func BenchmarkGokitJSONNegative(b *testing.B) {
-	stream := &blackholeStream{}
+	stream := &synchronizedStream{}
 	logger := log.NewContext(log.NewJSONLogger(stream)).With("ts", log.DefaultTimestampUTC)
 	lvllog := newLeveledLogger(logger, false)
 
@@ -63,7 +77,7 @@ func BenchmarkGokitJSONNegative(b *testing.B) {
 }
 
 func BenchmarkGokitTextPositive(b *testing.B) {
-	stream := &blackholeStream{}
+	stream := &synchronizedStream{}
 	logger := log.NewContext(log.NewLogfmtLogger(stream)).With("ts", log.DefaultTimestampUTC)
 	lvllog := newLeveledLogger(logger, true)
 
@@ -81,7 +95,7 @@ func BenchmarkGokitTextPositive(b *testing.B) {
 }
 
 func BenchmarkGokitTextNegative(b *testing.B) {
-	stream := &blackholeStream{}
+	stream := &synchronizedStream{}
 	logger := log.NewContext(log.NewLogfmtLogger(stream)).With("ts", log.DefaultTimestampUTC)
 	lvllog := newLeveledLogger(logger, false)
 
```

run 5 times on each version, [benchstat](https://godoc.org/rsc.io/benchstat) shows this change (allocs elided):

```
name                 old time/op  new time/op  delta
GokitJSONPositive-4  4.60µs ± 6%  4.68µs ± 3%    ~     (p=0.310 n=5+5)
GokitJSONNegative-4   334ns ± 2%   331ns ± 3%    ~     (p=0.302 n=5+5)
GokitTextPositive-4  1.60µs ± 2%  1.59µs ± 2%    ~     (p=0.889 n=5+5)
GokitTextNegative-4   128ns ± 8%   118ns ± 2%  -7.97%  (p=0.032 n=5+5)
```