@peterbourgon 2016-05-13
For the record, here is the benchstat with `-count=5` for all loggers.

```
GokitJSONPositive-4      5.13µs ±16%   1.33kB ± 0%   36.0 ± 0%
Log15JSONPositive-4      12.6µs ±20%   1.80kB ± 0%   40.0 ± 0%
LogrusJSONPositive-4     6.15µs ±20%   2.35kB ± 0%   43.0 ± 0%

GokitJSONNegative-4       377ns ±16%     232B ± 0%   9.00 ± 0%
Log15JSONNegative-4       909ns ± 5%     392B ± 0%   9.00 ± 0%
LogrusJSONNegative-4     1.14µs ±24%     832B ± 0%   10.0 ± 0%

GokitTextPositive-4      1.65µs ± 4%     800B ± 0%   10.0 ± 0%
GologgingTextPositive-4  2.11µs ± 6%     952B ± 0%   17.0 ± 0%
Log15TextPositive-4      8.05µs ±11%   1.12kB ± 0%   24.0 ± 0%
LogrusTextPositive-4     2.79µs ± 9%     880B ± 0%   17.0 ± 0%
SeelogTextPositive-4     4.92µs ± 8%     456B ± 0%   12.0 ± 0%

GokitTextNegative-4       123ns ± 7%    64.0B ± 0%   3.00 ± 0%
GologgingTextNegative-4   215ns ± 8%     160B ± 0%   3.00 ± 0%
Log15TextNegative-4       542ns ±26%     128B ± 0%   1.00 ± 0%
LogrusTextNegative-4     42.2ns ± 2%    16.0B ± 0%   1.00 ± 0%
SeelogTextNegative-4      132ns ±13%    64.0B ± 0%   3.00 ± 0%
```