@ChrisHines 2016-05-13
Cool, but wow, there is a lot of noise in those measurements.