@peterbourgon 2016-05-13
@imkira Ready to merge, from my perspective.